Add user-selectable currency display and preference
Implement user currency preference with new endpoints for getting and updating currency settings. Add UI toggles and components to display balances, transactions, and inputs in either YCP or the user's chosen currency. Include a new Currency Settings page, update AppState to track currency info, and ensure conversions and error handling throughout the app. All transactions remain in YCP; display adapts to user preference.

+++ Promise.Lib/Models/Entities.cs
@@ -50,7 +50,7 @@ public class PromiseTransaction
 public class Rate
 {
     public byte CurrencyId { get; set; }
-    public float AmountFor100 { get; set; }
+    public double AmountFor100 { get; set; }
     public DateTime UpdateDate { get; set; }
 }
